requirements.txt
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
autoflake==2.3.1
backcall==0.2.0
beautifulsoup4==4.13.4
black==25.1.0
bleach==6.2.0
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
docopt==0.6.2
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
greenlet==3.2.2
identify==2.6.10
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipython==8.12.3
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nodeenv==1.9.1
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.8
playwright==1.52.0
pluggy==1.6.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodome==3.23.0
pyee==13.0.0
pyflakes==3.3.2
Pygments==2.19.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
text-unidecode==1.3
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.31.2
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.21.0
